Move math utility methods to the sharedmath module ,format documentation
