--- src/main/java/dev/codesoapbox/dummy4j/dummies/color/Round.java
@@ -1,14 +1,0 @@
-package dev.codesoapbox.dummy4j.dummies.color;
-
-/**
- * @since 0.4.0
- */
-final class Round {
-
-    private Round() {
-    }
-
-    public static float toTwoDecimals(float value) {
-        return (float) (Math.round(value * 100.0F) / 100.0);
-    }
-}
