+++ src/main/java/dev/codesoapbox/dummy4j/dummies/sharedmath/NumberFormatter.java
@@ -1,21 +1,29 @@
-package dev.codesoapbox.dummy4j.dummies.color;
+package dev.codesoapbox.dummy4j.dummies.sharedmath;
 
 import java.text.DecimalFormat;
 
 /**
- * @since 0.4.0
+ * Provides convenience methods for number formatting
+ *
+ * @since SNAPSHOT
  */
-final class NumberFormatter {
+public final class NumberFormatter {
 
     private static final DecimalFormat DECIMAL = new DecimalFormat("0.##");
 
     private NumberFormatter() {
     }
 
+    /**
+     * Formats the given value to two decimal places, without trailing zeros
+     */
     public static String toTwoDecimals(float value) {
         return DECIMAL.format(value);
     }
 
+    /**
+     * Formats the given value to a percent value
+     */
     public static String toPercent(float value) {
         return DECIMAL.format(value * 100F) + "%";
     }
