+++ src/main/java/dev/codesoapbox/dummy4j/dummies/sharedmath/Round.java
@@ -1,0 +1,19 @@
+package dev.codesoapbox.dummy4j.dummies.sharedmath;
+
+/**
+ * Provides convenience methods for rounding
+ *
+ * @since SNAPSHOT
+ */
+public final class Round {
+
+    private Round() {
+    }
+
+    /**
+     * Rounds the given value to two decimal places
+     */
+    public static float toTwoDecimals(float value) {
+        return (float) (Math.round(value * 100.0F) / 100.0);
+    }
+}
