+++ src/test/java/dev/codesoapbox/dummy4j/dummies/sharedmath/RoundTest.java
@@ -1,4 +1,4 @@
-package dev.codesoapbox.dummy4j.dummies.color;
+package dev.codesoapbox.dummy4j.dummies.sharedmath;
 
 import org.junit.jupiter.api.Test;
